# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FixYa - Plataforma de servicios técnicos para el hogar
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /requests:
    get:
      operationId: listRequests
      tags: [requests]
      summary: List service requests (filtered by role)
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pendiente, aceptada, en_progreso, completada, cancelada, rechazada]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequestList"
    post:
      operationId: createRequest
      tags: [requests]
      summary: Create a service request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceRequestInput"
      responses:
        "201":
          description: Created request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}:
    get:
      operationId: getRequest
      tags: [requests]
      summary: Get service request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /requests/{id}/accept:
    post:
      operationId: acceptRequest
      tags: [requests]
      summary: Accept a service request (technician)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}/reject:
    post:
      operationId: rejectRequest
      tags: [requests]
      summary: Reject a service request (technician)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectInput"
      responses:
        "200":
          description: Request rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}/complete:
    post:
      operationId: completeRequest
      tags: [requests]
      summary: Mark a service request as completed (technician)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"
  /requests/{id}/cancel:
    post:
      operationId: cancelRequest
      tags: [requests]
      summary: Cancel a service request (user)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Request cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceRequest"

  # ─── PAYMENTS ─────────────────────────────────────────────────────────────────
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews
      parameters:
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{id}:
    get:
      operationId: getReview
      tags: [reviews]
      summary: Get review by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ─── DASHBOARD ────────────────────────────────────────────────────────────────
...
components:
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # ─── USER ──────────────────────────────────────────────────────────────────
    User:
      type: object
      properties:
        id:
          type: integer
        clerkId:
          type: string
        nombre:
          type: string
        apellido:
          type: string
        email:
          type: string
        telefono:
          type: ["string", "null"]
        direccion:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [usuario, tecnico, administrador]
        activo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkId
        - nombre
        - apellido
        - email
        - role
        - activo
        - createdAt
...
    Service:
      type: object
      properties:
        id:
          type: integer
        nombre:
          type: string
        descripcion:
          type: ["string", "null"]
        icono:
          type: ["string", "null"]
        precioBase:
          type: number
        activo:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - nombre
        - precioBase
        - activo
        - createdAt
...
    Technician:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        serviceId:
          type: integer
        especialidad:
          type: string
        descripcion:
          type: ["string", "null"]
        experienciaAnios:
          type: integer
        precioHora:
          type: number
        disponible:
          type: boolean
        status:
          type: string
          enum: [pendiente, aprobado, rechazado, suspendido]
        promedioCalificacion:
          type: number
        totalTrabajos:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        service:
          $ref: "#/components/schemas/Service"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - serviceId
        - especialidad
        - experienciaAnios
        - precioHora
        - disponible
        - status
        - promedioCalificacion
        - totalTrabajos
        - createdAt
...
    RejectInput:
      type: object
      properties:
        motivo:
          type: string
      required:
        - motivo

    # ─── SERVICE REQUEST ───────────────────────────────────────────────────────
    ServiceRequest:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        technicianId:
          type: ["integer", "null"]
        serviceId:
          type: integer
        descripcion:
          type: string
        direccion:
          type: string
        fechaServicio:
          type: string
          format: date-time
        status:
          type: string
          enum: [pendiente, aceptada, en_progreso, completada, cancelada, rechazada]
        precioAcordado:
          type: ["number", "null"]
        motivoCancelacion:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        technician:
          $ref: "#/components/schemas/Technician"
        service:
          $ref: "#/components/schemas/Service"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - serviceId
        - descripcion
        - direccion
        - fechaServicio
        - status
        - createdAt
    ServiceRequestList:
      type: object
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/ServiceRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - requests
        - total
        - page
        - limit
    ServiceRequestInput:
      type: object
      properties:
        serviceId:
          type: integer
        technicianId:
          type: integer
        descripcion:
          type: string
        direccion:
          type: string
        fechaServicio:
          type: string
          format: date-time
        precioAcordado:
          type: number
      required:
        - serviceId
        - descripcion
        - direccion
        - fechaServicio

    # ─── PAYMENT ───────────────────────────────────────────────────────────────
...
    Review:
      type: object
      properties:
        id:
          type: integer
        requestId:
          type: integer
        userId:
          type: integer
        technicianId:
          type: integer
        calificacion:
          type: integer
        comentario:
          type: ["string", "null"]
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - requestId
        - userId
        - technicianId
        - calificacion
        - createdAt
    ReviewList:
      type: object
      properties:
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - reviews
        - total
        - page
        - limit
    ReviewInput:
      type: object
      properties:
        requestId:
          type: integer
        technicianId:
          type: integer
        calificacion:
          type: integer
          minimum: 1
          maximum: 5
        comentario:
          type: string
      required:
        - requestId
        - technicianId
        - calificacion

    # ─── DASHBOARD ─────────────────────────────────────────────────────────────
...