# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * FixYa - Plataforma de servicios técnicos para el hogar
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  nombre: string;
  /** @nullable */
  descripcion?: string | null;
  /** @nullable */
  icono?: string | null;
  precioBase: number;
  activo: boolean;
  createdAt: Date;
}